+++ src/pages/index.tsx
@@ -1,9 +1,11 @@
 import React, { useState } from "react"
-import { graphql, useStaticQuery, Node, Link } from "gatsby"
-import styled, { css, DefaultTheme } from "styled-components"
+import { graphql, useStaticQuery, Node } from "gatsby"
+import styled, { DefaultTheme } from "styled-components"
 
 import SEO from "../components/seo"
-import { getImage, IGatsbyImageData, GatsbyImage } from "gatsby-plugin-image"
+import { getImage, IGatsbyImageData } from "gatsby-plugin-image"
+import { ExternalLink, InternalLink } from "../components/links"
+import Card from "../components/portfolio-card"
 
 interface IndexQueryData {
   allFile: {
@@ -16,8 +18,8 @@ interface IndexQueryData {
   }
 }
 
-type CardType = "Case Study" | "Research" | "Project"
-type ShowType = CardType | "All"
+export type CardType = "Case Study" | "Research" | "Project"
+export type ShowType = CardType | "All"
 
 const IndexPage = (): JSX.Element => {
   const [show, setShow] = useState<ShowType>("All")
@@ -117,141 +119,6 @@ const Description = styled.div`
   }
 `
 
-const linkStyle = css`
-  color: inherit;
-  text-decoration: underline;
-  text-decoration-thickness: 1px;
-`
-
-const InternalLink = styled(Link)`
-  ${linkStyle}
-`
-
-const ExternalLink = styled.a`
-  ${linkStyle}
-`
-
-const Cards = styled.div`
-  display: grid;
-  grid-auto-rows: 1fr;
-  grid-auto-flow: row dense;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 3rem 1.5rem;
-  margin: 3rem 0 0 0;
-
-  @media screen and (max-width: 50rem) {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 2rem 1rem;
-  }
-`
-
-interface CardProps {
-  title: string
-  link?: string
-  type: CardType
-  image?: IGatsbyImageData
-  active?: boolean
-  width: CardWidth
-  background?: boolean
-}
-
-const Card = ({
-  title,
-  link,
-  type,
-  image,
-  width,
-  active = true,
-  background = false,
-}: CardProps) => {
-  const content = (
-    <>
-      {image && (
-        <CardImageWrapper background={background}>
-          <CardImage background={background} alt={title} image={image} />
-        </CardImageWrapper>
-      )}
-      <CardTitle>{title}</CardTitle>
-      <CardSubtitle>{type}</CardSubtitle>
-    </>
-  )
-
-  return (
-    <CardContainer active={active} width={width}>
-      {link ? <CardLink href={link}>{content}</CardLink> : content}
-    </CardContainer>
-  )
-}
-
-const CardImageWrapper = styled.div`
-  display: flex;
-  align-items: stretch;
-  width: 100%;
-  height: 100%;
-  min-height: 10rem;
-  opacity: 0.7;
-  transition: opacity 0.5s;
-
-  &:hover {
-    opacity: 1;
-  }
-
-  ${({ background }: { background: boolean }) =>
-    background &&
-    css`
-      background: #eeeeee;
-    `}
-`
-
-interface CardImageProps {
-  background: boolean
-}
-
-const CardImage = styled(GatsbyImage)`
-  ${({ background }: CardImageProps) =>
-    background &&
-    `
-    width: 50%;
-    margin: auto;
-    box-shadow: 0 0 1rem 0.25rem #00000022;
-  `}
-`
-
-const CardLink = styled.a`
-  display: flex;
-  flex-direction: column;
-  flex-grow: 0;
-  height: 100%;
-  color: ${({ theme }) => theme.foreground.default};
-  text-decoration: none;
-`
-
-type CardWidth = 1 | 2 | 3 | 4
-
-interface CardContainerProps {
-  width: CardWidth
-  active: boolean
-}
-
-const CardContainer = styled.div<CardContainerProps>`
-  display: flex;
-  flex-direction: column;
-  grid-column: span ${({ width }) => width};
-  line-height: 1.3;
-  opacity: ${({ active }) => (active ? 1 : 0.3)};
-  filter: ${({ active }) => (active ? "none" : "blur(4px)")};
-  transition: opacity 0.5s, filter 0.5s;
-  pointer-events: ${({ active }) => (active ? "default" : "none")};
-`
-
-const CardTitle = styled.span`
-  margin: 0.5rem 0 0 0;
-`
-
-const CardSubtitle = styled.span`
-  color: ${({ theme }) => theme.foreground.sub};
-`
-
 const Container = styled.div`
   margin: 6rem 0 0 0;
   line-height: 1.5;
@@ -270,6 +137,20 @@ const Bio = styled.div`
   @media screen and (max-width: 50rem) {
     grid-template-columns: 1fr 1fr;
     gap: 1rem;
+  }
+`
+
+const Cards = styled.div`
+  display: grid;
+  grid-auto-rows: 1fr;
+  grid-auto-flow: row dense;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 3rem 1.5rem;
+  margin: 3rem 0 0 0;
+
+  @media screen and (max-width: 50rem) {
+    grid-template-columns: repeat(2, 1fr);
+    gap: 2rem 1rem;
   }
 `
